pyproject.toml
[tool.poetry]
name = "boni-py"
version = "0.1.0"
description = ""
authors = ["Wolfhound905 <aiden8green@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.20.0"
beanie = "^1.11.0"
orjson = "^3.8.2"
interactions-py = "^5.6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
